Create new Cards.h header for data structures Fix missing namespaces
